requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
pip==23.2.1
sqlparse==0.5.3
tzdata==2024.2
rest-framework-simplejwt==0.0.2
